+++ Source/Private/_ASPendingState.mm
@@ -73,8 +73,11 @@ typedef struct {
   int setEdgeAntialiasingMask:1;
   int setIsAccessibilityElement:1;
   int setAccessibilityLabel:1;
+  int setAccessibilityAttributedLabel:1;
   int setAccessibilityHint:1;
+  int setAccessibilityAttributedHint:1;
   int setAccessibilityValue:1;
+  int setAccessibilityAttributedValue:1;
   int setAccessibilityTraits:1;
   int setAccessibilityFrame:1;
   int setAccessibilityLanguage:1;
@@ -121,8 +124,11 @@ typedef struct {
   BOOL asyncTransactionContainer;
   BOOL isAccessibilityElement;
   NSString *accessibilityLabel;
+  NSAttributedString *accessibilityAttributedLabel;
   NSString *accessibilityHint;
+  NSAttributedString *accessibilityAttributedHint;
   NSString *accessibilityValue;
+  NSAttributedString *accessibilityAttributedValue;
   UIAccessibilityTraits accessibilityTraits;
   CGRect accessibilityFrame;
   NSString *accessibilityLanguage;
@@ -271,8 +277,11 @@ static BOOL defaultAllowsEdgeAntialiasing = NO;
   borderColor = blackColorRef;
   isAccessibilityElement = NO;
   accessibilityLabel = nil;
+  accessibilityAttributedLabel = nil;
   accessibilityHint = nil;
+  accessibilityAttributedHint = nil;
   accessibilityValue = nil;
+  accessibilityAttributedValue = nil;
   accessibilityTraits = UIAccessibilityTraitNone;
   accessibilityFrame = CGRectZero;
   accessibilityLanguage = nil;
@@ -586,9 +595,26 @@ static BOOL defaultAllowsEdgeAntialiasing = NO;
 
 - (void)setAccessibilityLabel:(NSString *)newAccessibilityLabel
 {
-  _flags.setAccessibilityLabel = YES;
-  if (accessibilityLabel != newAccessibilityLabel) {
+  if (![accessibilityLabel isEqualToString:newAccessibilityLabel]) {
+    _flags.setAccessibilityLabel = YES;
+    _flags.setAccessibilityAttributedLabel = YES;
     accessibilityLabel = [newAccessibilityLabel copy];
+    accessibilityAttributedLabel = [[NSAttributedString alloc] initWithString:newAccessibilityLabel];
+  }
+}
+
+- (NSAttributedString *)accessibilityAttributedLabel
+{
+  return accessibilityAttributedLabel;
+}
+
+- (void)setAccessibilityAttributedLabel:(NSAttributedString *)newAccessibilityAttributedLabel
+{
+  if (![accessibilityAttributedLabel isEqualToAttributedString: newAccessibilityAttributedLabel]) {
+    _flags.setAccessibilityAttributedLabel = YES;
+    _flags.setAccessibilityLabel = YES;
+    accessibilityAttributedLabel = [newAccessibilityAttributedLabel copy];
+    accessibilityLabel = [newAccessibilityAttributedLabel.string copy];
   }
 }
 
@@ -599,8 +625,27 @@ static BOOL defaultAllowsEdgeAntialiasing = NO;
 
 - (void)setAccessibilityHint:(NSString *)newAccessibilityHint
 {
-  _flags.setAccessibilityHint = YES;
-  accessibilityHint = [newAccessibilityHint copy];
+  if (![accessibilityHint isEqualToString:newAccessibilityHint]) {
+    _flags.setAccessibilityHint = YES;
+    _flags.setAccessibilityAttributedHint = YES;
+    accessibilityHint = [newAccessibilityHint copy];
+    accessibilityAttributedHint = [[NSAttributedString alloc] initWithString:newAccessibilityHint];
+  }
+}
+
+- (NSAttributedString *)accessibilityAttributedHint
+{
+  return accessibilityAttributedHint;
+}
+
+- (void)setAccessibilityAttributedHint:(NSAttributedString *)newAccessibilityAttributedHint
+{
+  if (![accessibilityAttributedHint isEqual:newAccessibilityAttributedHint]) {
+    _flags.setAccessibilityAttributedHint = YES;
+    _flags.setAccessibilityHint = YES;
+    accessibilityAttributedHint = [newAccessibilityAttributedHint copy];
+    accessibilityHint = [newAccessibilityAttributedHint.string copy];
+  }
 }
 
 - (NSString *)accessibilityValue
@@ -610,8 +655,27 @@ static BOOL defaultAllowsEdgeAntialiasing = NO;
 
 - (void)setAccessibilityValue:(NSString *)newAccessibilityValue
 {
-  _flags.setAccessibilityValue = YES;
-  accessibilityValue = [newAccessibilityValue copy];
+  if (![accessibilityValue isEqualToString:newAccessibilityValue]) {
+    _flags.setAccessibilityValue = YES;
+    _flags.setAccessibilityAttributedValue = YES;
+    accessibilityValue = [newAccessibilityValue copy];
+    accessibilityAttributedValue = [[NSAttributedString alloc] initWithString:newAccessibilityValue];
+  }
+}
+
+- (NSAttributedString *)accessibilityAttributedValue
+{
+  return accessibilityAttributedValue;
+}
+
+- (void)setAccessibilityAttributedValue:(NSAttributedString *)newAccessibilityAttributedValue
+{
+  if (![accessibilityAttributedValue isEqualToAttributedString:newAccessibilityAttributedValue]) {
+    _flags.setAccessibilityAttributedValue = YES;
+    _flags.setAccessibilityValue = YES;
+    accessibilityAttributedValue = [newAccessibilityAttributedValue copy];
+    accessibilityValue = [newAccessibilityAttributedValue.string copy];
+  }
 }
 
 - (UIAccessibilityTraits)accessibilityTraits
@@ -994,11 +1058,20 @@ static BOOL defaultAllowsEdgeAntialiasing = NO;
   if (flags.setAccessibilityLabel)
     view.accessibilityLabel = accessibilityLabel;
 
+  if (AS_AT_LEAST_IOS11 && flags.setAccessibilityAttributedLabel)
+    [view setValue:accessibilityAttributedLabel forKey:@"accessibilityAttributedLabel"];
+
   if (flags.setAccessibilityHint)
     view.accessibilityHint = accessibilityHint;
 
+  if (AS_AT_LEAST_IOS11 && flags.setAccessibilityAttributedHint)
+    [view setValue:accessibilityAttributedHint forKey:@"accessibilityAttributedHint"];
+
   if (flags.setAccessibilityValue)
     view.accessibilityValue = accessibilityValue;
+
+  if (AS_AT_LEAST_IOS11 && flags.setAccessibilityAttributedValue)
+    [view setValue:accessibilityAttributedValue forKey:@"accessibilityAttributedValue"];
 
   if (flags.setAccessibilityTraits)
     view.accessibilityTraits = accessibilityTraits;
@@ -1142,6 +1215,11 @@ static BOOL defaultAllowsEdgeAntialiasing = NO;
   pendingState.accessibilityLabel = view.accessibilityLabel;
   pendingState.accessibilityHint = view.accessibilityHint;
   pendingState.accessibilityValue = view.accessibilityValue;
+  if (AS_AT_LEAST_IOS11) {
+    pendingState.accessibilityAttributedLabel = [view valueForKey: @"accessibilityAttributedLabel"];
+    pendingState.accessibilityAttributedHint = [view valueForKey: @"accessibilityAttributedHint"];
+    pendingState.accessibilityAttributedValue = [view valueForKey: @"accessibilityAttributedValue"];
+  }
   pendingState.accessibilityTraits = view.accessibilityTraits;
   pendingState.accessibilityFrame = view.accessibilityFrame;
   pendingState.accessibilityLanguage = view.accessibilityLanguage;
@@ -1219,8 +1297,11 @@ static BOOL defaultAllowsEdgeAntialiasing = NO;
   || flags.setSemanticContentAttribute
   || flags.setIsAccessibilityElement
   || flags.setAccessibilityLabel
+  || flags.setAccessibilityAttributedLabel
   || flags.setAccessibilityHint
+  || flags.setAccessibilityAttributedHint
   || flags.setAccessibilityValue
+  || flags.setAccessibilityAttributedValue
   || flags.setAccessibilityTraits
   || flags.setAccessibilityFrame
   || flags.setAccessibilityLanguage